requirements.txt
psycopg==3.2.8
psycopg-binary==3.2.8
typing-extensions==4.13.2
uv==0.7.3
asgiref==3.8.1
django==4.2
sqlparse==0.5.3
uvloop==0.21.0
adrf==0.1.9
async-property==0.2.2
djangorestframework==3.16.0
dotenv==0.9.9
python-dotenv==1.1.0
click==8.2.0
h11==0.16.0
uvicorn==0.34.2
drf-yasg==1.21.10
inflection==0.5.1
packaging==25.0
pytz==2025.2
pyyaml==6.0.2
uritemplate==4.1.1
django-filter==25.1
attrs==25.3.0
drf-spectacular==0.28.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
referencing==0.36.2
rpds-py==0.24.0
unidecode==1.4.0
django-redis==5.4.0
redis==6.1.0
